refactor(#3038): move turn bridge status-panel helpers
